Move Rust event copying into core
